Extract SourceStruct into its own file
