+++ lib/flight_configuration/source_struct.rb
@@ -25,14 +25,43 @@
 # https://github.com/openflighthpc/flight_configuration
 #==============================================================================
 
-require "flight_configuration/version"
-require "flight_configuration/deep_stringify_keys"
-require "flight_configuration/logs"
-require "flight_configuration/base_dsl"
-require "flight_configuration/dsl"
-require "flight_configuration/source_struct"
-require "flight_configuration/rack_dsl"
-
 module FlightConfiguration
-  class Error < StandardError; end
+  # Stores a reference to where a particular key came from
+  # NOTE: The type specifies if it came from the :env or :file
+  SourceStruct = Struct.new(:key, :source, :type, :value, :config) do
+    def attribute
+      @attribute ||= config.class.attributes[key] || {}
+    end
+
+    def transformable?
+      transformed_value if @transformable.nil?
+      @transformable
+    end
+
+    def recognized?
+      !attribute.empty?
+    end
+
+    def transformed_value
+      return @transformed_value unless @transformable.nil?
+      @transformable = true
+      transform = attribute[:transform]
+      @transformed_value = if transform.nil?
+        value
+      elsif transform.respond_to?(:call)
+        transform.call(value)
+      else
+        value.send(transform)
+      end
+    rescue
+      # # NOTE: Ideally the error would be logged, however this can't be done
+      # #       without forming a recursive loop
+      # if active_errors?
+      #   config.errors.add(key.to_sym, type: :transform, message: 'failed to coerce the data type')
+      # else
+      #   raise Error, "Failed to coerce attribute: #{key}"
+      # end
+      @transformable = false
+    end
+  end
 end
